+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/DependencyGraphNodeCache.kt
@@ -24,6 +24,7 @@ import dev.zacsweers.metro.compiler.tracing.Tracer
 import dev.zacsweers.metro.compiler.tracing.traceNested
 import org.jetbrains.kotlin.ir.builders.irCall
 import org.jetbrains.kotlin.ir.declarations.IrClass
+import org.jetbrains.kotlin.ir.declarations.IrDeclaration
 import org.jetbrains.kotlin.ir.declarations.IrDeclarationWithName
 import org.jetbrains.kotlin.ir.declarations.IrFunction
 import org.jetbrains.kotlin.ir.declarations.IrOverridableDeclaration
@@ -110,6 +111,7 @@ internal class DependencyGraphNodeCache(
     private val bindsFunctions = mutableListOf<Pair<MetroSimpleFunction, IrContextualTypeKey>>()
     private val bindsCallables = mutableSetOf<BindsCallable>()
     private val multibindsCallables = mutableSetOf<MultibindsCallable>()
+    private val optionalKeys = mutableMapOf<IrTypeKey, MutableSet<BindsOptionalOfCallable>>()
     private val scopes = mutableSetOf<IrAnnotation>()
     private val providerFactories = mutableListOf<Pair<IrTypeKey, ProviderFactory>>()
     private val extendedGraphNodes = mutableMapOf<IrTypeKey, DependencyGraphNode>()
@@ -755,6 +757,9 @@ internal class DependencyGraphNodeCache(
         container.bindsMirror?.let { bindsMirror ->
           bindsCallables += bindsMirror.bindsCallables
           multibindsCallables += bindsMirror.multibindsCallables
+          for (callable in bindsMirror.optionalKeys) {
+            optionalKeys.getOrPut(callable.typeKey) { mutableSetOf() } += callable
+          }
         }
 
         // Record an IC lookup of the container class
@@ -776,6 +781,7 @@ internal class DependencyGraphNodeCache(
           bindsCallables = bindsCallables,
           bindsFunctions = bindsFunctions.map { it.first },
           multibindsCallables = multibindsCallables,
+          optionalKeys = optionalKeys,
           providerFactories = providerFactories,
           accessors = accessors,
           injectors = injectors,
@@ -870,10 +876,13 @@ internal class DependencyGraphNodeCache(
             bindingContainer.bindsMirror?.let { bindsMirror ->
               bindsCallables += bindsMirror.bindsCallables
               multibindsCallables += bindsMirror.multibindsCallables
+              for (callable in bindsMirror.optionalKeys) {
+                optionalKeys.getOrPut(callable.typeKey) { mutableSetOf() } += callable
+              }
             }
           }
         }
-      } else if (!isGraph) {
+      } else {
         providerFactories +=
           bindingContainerTransformer.factoryClassesFor(metroGraph ?: graphDeclaration)
       }
@@ -890,6 +899,7 @@ internal class DependencyGraphNodeCache(
           accessors = accessors,
           bindsCallables = bindsCallables,
           multibindsCallables = multibindsCallables,
+          optionalKeys = optionalKeys,
           isExternal = true,
           proto = null,
           extendedGraphNodes = extendedGraphNodes,